requirements.txt
django == 4.1.13
djangorestframework == 3.14.0
django-allauth == 0.61.1
pillow == 10.3.0
gunicorn == 20.1.0
dj-database-url==0.5.0
psycopg2==2.9.3
boto3==1.34.87
django-storages==1.14.0
botocore==1.34.87
six==1.16.0
jmespath==1.0.1
python-dateutil==2.9.0
s3transfer==0.10.1
django-cors-headers==4.3.1